2-Aside/wallet-service/requirements.txt
# Wallet Service Dependencies
azure-functions==1.18.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9
sqlalchemy==2.0.27
pymssql==2.2.11
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pytz==2024.1
httpx==0.26.0
